scale-down-gke-cluster/requirements.txt
functions-framework==3.*
google-cloud-container
google-cloud